WebScraping-BS4/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.4
exceptiongroup==1.0.4
fake-useragent==1.1.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-speech==2.16.2
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httmock==1.4.0
http-request-randomizer==1.3.2
idna==3.4
importlib-resources==5.10.0
outcome==1.2.0
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.10
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0
zipp==3.11.0
